tried to figure out isHitAt did not get it to work got rid of errors though. Also refactored TextTool.ts to Text.ts
