+++ src/plottables/Text.ts
@@ -16,15 +16,15 @@ import {
 import { ValueDisplayMode } from "@/types";
 //import Two from "two.js";
 import { Vector } from "two.js/src/vector";
-import { Text } from "two.js/src/text";
+import { Text as TwoJsText } from "two.js/src/text";
 import { Group } from "two.js/src/group";
 
-//had to name file TextTool so that it does not conflict wit two.js/src/text
-export default class TextTool extends Nodule {
-  public text: Text;
+//had to name file Text so that it does not conflict wit two.js/src/text
+export default class Text extends Nodule {
+  public text: TwoJsText;
   constructor(txt: string, x: number, y: number, noduleName: string = "None") {
     super(noduleName);
-    this.text = new Text(txt, x, y);
+    this.text = new TwoJsText(txt, x, y);
   }
   /**
    * The vector location of the Label on the default unit sphere
@@ -50,7 +50,7 @@ export default class TextTool extends Nodule {
    * @param factor The ratio of the old magnification factor over the new magnification factor
    */
   static updateTextScaleFactorForZoom(factor: number): void {
-    TextTool.textScaleFactor *= factor;
+    Text.textScaleFactor *= factor;
   }
 
   static isEarthMode = false;
@@ -85,7 +85,7 @@ export default class TextTool extends Nodule {
   }
   updateDisplay(): void {
     this.normalDisplay();
-    console.debug("Calling TextTool.normalDisplay();");
+    console.debug("Calling Text.normalDisplay();");
     /**None**/
   }
   toSVG(nonScaling?: { stroke: boolean; text: boolean; pointRadius: boolean; scaleFactor: number; }, svgForIcon?: boolean): toSVGType[] {
@@ -111,5 +111,28 @@ export default class TextTool extends Nodule {
       top: rect.top,
       width: rect.width
     };
+  }
+
+  set positionVector(idealUnitSphereVectorLocation: Vector3) {
+    this._locationVector
+      .copy(idealUnitSphereVectorLocation)
+      .multiplyScalar(SETTINGS.boundaryCircle.radius);
+    // Translate the whole group (i.e. all points front/back/glowing/drawn) to the new center vector
+    this.defaultScreenVectorLocation.set(
+      this._locationVector.x,
+      this._locationVector.y
+    );
+    //this.updateDisplay();  //<--- do not do this! disconnect the setting of position with the display, if you leave this in
+    //then this turns on the display of the vertex point of the angle marker in a bad way. It turns on the
+    //     // the display so that the following problem occurs.
+    //     //   1. Create/Measure an angle from three new points
+    //     //   2. Hide the point at the vertex
+    //     //   3. Create an angle bisector
+    //     //   4. Notice that the vertex point appears but is now not selectable (because the this.showing is false, but the
+    //     //    actual display is showing, so it is not found in the Highligher.ts handler subclass )
+    //     // I spent at least 8 hours looking for how this occurs ... :-()
+  }
+  get positionVector(): Vector3 {
+    return this._locationVector;
   }
 }
